+++ .dockerignore
@@ -1,0 +1,8 @@
+/node_modules
+/out
+/.next
+
+yarn.lock
+package-lock.json
+
+.env
