--- files/Imagens/_credits
@@ -1,25 +1,0 @@
-
-Many thanks to the following artists for sending their artwork to Linux Mint and/or for distributing it under Free licenses.
-
--------------------------------------------------------------------------------------------------------
-| Prefix       | Artist                   | Link                                   | License          |
--------------------------------------------------------------------------------------------------------
-| aburden      | Aaron Burden             | http://aaronburden.com                 | Unsplash License |
-| akrivec      | Ales Krivec              | https://dreamypixel.com                | Unsplash License |
-| dmirlea      | Daniel Mirlea            | https://danielmirlea.com               | Unsplash License |
-| jhuber       | Jan Huber                | https://www.jan-huber.ch               | Unsplash License |
-| jjforestbird | Joel & Jasmin Førestbird | https://theforestbirds.com             | Unsplash License |
-| ktaira       | Kazuki Taira             | https://unsplash.com/@kazuuuki_taira   | Unsplash License |
-| lherrera     | Lorenzo Herrera          | https://lorenzoherrera.com             | Unsplash License |
-| lvillasmil   | Luis Villasmil           | https://www.behance.net/luisv18        | Unsplash License |
-| mnohassi     | Mohamed Nohassi          | https://nohassi.bio.link               | Unsplash License |
-| mpiwnicki    | Marek Piwnicki           | https://marpiwnicki.github.io          | Unsplash License |
-| msaeling     | Max Saeling              | https://maxsaeling.de                  | Unsplash License |
-| navi         | Navi                     | https://unsplash.com/@navi_photography | Unsplash License |
-| pcarmona     | Paul Carmona             | https://www.sydneylens.com.au          | Unsplash License |
-| pczerwinski  | Pawel Czerwinski         | https://unsplash.com/@pawel_czerwinski | Unsplash License |
-| pfore        | Patrick Fore             | https://www.patrickfore.com            | Unsplash License |
-| sjohnson     | Steve Johnson            | https://artbystevej.com                | Unsplash License |
-| ssinclair    | Sean Sinclair            | https://unsplash.com/@seanwsinclair    | Unsplash License |
-| vbkremer     | Valentin B. Kremer       | https://www.vbk-media.net              | Unsplash License |
--------------------------------------------------------------------------------------------------------
